fix for testing code

+++ eclipse_ws/biobank2.tests/src/test/ualberta/med/biobank/internal/DbHelper.java
@@ -19,44 +19,54 @@ public class DbHelper {
         DbHelper.appService = appService;
     }
 
-    public static void removeContainers(List<ContainerWrapper> containers)
+    public static void deleteContainers(List<ContainerWrapper> containers)
         throws Exception {
+        if ((containers == null) || (containers.size() == 0))
+            return;
+
         for (ContainerWrapper container : containers) {
             if (container.hasChildren()) {
-                removeContainers(container.getChildren());
+                deleteContainers(container.getChildren());
             } else {
-                removeFromList(container.getSamples());
+                deleteFromList(container.getSamples());
             }
             container.reload();
             container.delete();
         }
     }
 
-    public static void removeStudies(List<StudyWrapper> studies)
+    public static void deleteStudies(List<StudyWrapper> studies)
         throws Exception {
+        if (studies == null)
+            return;
+
         for (StudyWrapper study : studies) {
-            removePatients(study.getPatientCollection());
+            deletePatients(study.getPatientCollection());
             study.reload();
             study.delete();
         }
     }
 
-    public static void removePatients(List<PatientWrapper> patients)
+    public static void deletePatients(List<PatientWrapper> patients)
         throws Exception {
+        if (patients == null)
+            return;
+
         for (PatientWrapper patient : patients) {
-            removeFromList(patient.getPatientVisitCollection());
+            deleteFromList(patient.getPatientVisitCollection());
             patient.reload();
             patient.delete();
         }
     }
 
-    public static void removeFromList(List<? extends ModelWrapper<?>> list)
+    public static void deleteFromList(List<? extends ModelWrapper<?>> list)
         throws Exception {
-        if (list != null) {
-            for (ModelWrapper<?> object : list) {
-                object.reload();
-                object.delete();
-            }
+        if (list == null)
+            return;
+
+        for (ModelWrapper<?> object : list) {
+            object.reload();
+            object.delete();
         }
     }
